chore(admin): 분석 지원자 성비 페이지 폴더 구조 변경
