+++ apps/admin/src/components/analysis/GenderRatio/FinalRoundPassed/FinalRoundPassed.tsx
@@ -1,16 +1,12 @@
 import { styled } from 'styled-components';
 import { flex } from '@maru/utils';
-import FinalRoundPassedScore from './FinalRoundPassedScore';
-import FinalRoundPassedTable from './FinalRoundPassedTable';
+import FinalRoundPassedMain from './FinalRoundPassedMain';
 import FinalRoundPassedDetail from './FinalRoundPassedDetail';
 
 const FinalRoundPassed = () => {
   return (
     <Layout>
-      <LeftBox>
-        <FinalRoundPassedScore />
-        <FinalRoundPassedTable />
-      </LeftBox>
+      <FinalRoundPassedMain />
       <FinalRoundPassedDetail />
     </Layout>
   );
@@ -25,10 +21,4 @@ const Layout = styled.div`
   })}
   width: 100%;
   gap: 75px;
-`;
-
-const LeftBox = styled.div`
-  ${flex({ flexDirection: 'column' })}
-  width: 100%;
-  gap: 100px;
 `;
